Update auth utility functions
