--- kairo-rest-feature/src/main/kotlin/kairo/rest/auth/deny.kt
@@ -1,10 +1,0 @@
-package kairo.rest.auth
-
-import kairo.rest.exception.EndpointAlwaysDenies
-
-/**
- * Denies all requests.
- */
-@Suppress("UnusedReceiverParameter")
-public fun Auth.deny(): Auth.Result =
-  Auth.Result.Exception(EndpointAlwaysDenies())
